Een mooie taak and kamer structuur toegepast in alle kamer, controller en taken
