+++ src/main/java/org/scrumEscape/classes/Kamers/DailyScrum.java
@@ -2,14 +2,16 @@ package org.scrumEscape.classes.Kamers;
 
 import org.scrumEscape.base.Kamer;
 import org.scrumEscape.classes.Monster;
-import org.scrumEscape.classes.Taak.Puzzel;
+import org.scrumEscape.classes.taak.Puzzel;
 import org.scrumEscape.interfaces.GameObserver;
 import org.scrumEscape.interfaces.TaakStrategie;
+
 import java.util.ArrayList;
 import java.util.Map;
+
 public class DailyScrum extends Kamer {
 
-	public DailyScrum( GameObserver gameObserver) {
+	public DailyScrum(GameObserver gameObserver) {
 		super("Daily Scrum", new Monster(), gameObserver);
 	}
 
@@ -21,11 +23,11 @@ public class DailyScrum extends Kamer {
 		Puzzel puzzel1 = new Puzzel(
 				"Verbind de teamleden bij de juiste taken",
 				Map.ofEntries(
-				Map.entry("Teun:Scrum Master", "Ik heb een probleem met het opstarten"),
-				Map.entry("Roody:Developer", "Ik heb gisteren een taak afgerond"),
-				Map.entry("Kyran:Developer", "Ik heb een andere taak afgerond"),
-				Map.entry("Andre:Product Owner", "Ik heb het afsluiten gemaakt")
-		));
+						Map.entry("Teun:Scrum Master", "Ik heb een probleem met het opstarten"),
+						Map.entry("Roody:Developer", "Ik heb gisteren een taak afgerond"),
+						Map.entry("Kyran:Developer", "Ik heb een andere taak afgerond"),
+						Map.entry("Andre:Product Owner", "Ik heb het afsluiten gemaakt")
+				));
 
 		opdrachten.add(puzzel1);
